requirements.txt
Pillow==7.0.0
scipy==1.4.1
SoundFile==0.10.3.post1